refactor(bridge): extract ProxyAuthHandler
